Move read and write descriptors to avoid circular imports
